Home Page and basic folder layout
Here is what I think the theme of the site should be, as well as what pages we should include in this website. Let me know what you think I should add. Keep note that this is a base for what we should do not what we have to do.